Выносит функции и данные в отдельные файлы
