+++ data.php
@@ -1,0 +1,10 @@
+<?php
+
+//Переменные
+$user_id = 1;
+$user_name = "Константин";
+
+// Показывать или нет выполненные задачи
+$show_complete_tasks = rand(0, 1);
+
+?>
